Major refactoring: bpm() method now returns double, tuned parameters fo SoundEnergyBPMDetector, removed unecessary println

+++ src/main/scala/me/mziccard/audio/bpm/BPMDetector.scala
@@ -8,6 +8,6 @@ trait BPMDetector {
    * Compute the beats per minute of an audio track
    * @return Track tempo in beats per minute
    **/
-  def bpm() : Int;
+  def bpm() : Double;
 
 }
